+++ scripts/benchmarks/benchmark_rsl_rl.py
@@ -166,15 +166,18 @@ def main(
     world_rank = 0
     world_size = 1
     if args_cli.distributed:
-        env_cfg.sim.device = f"cuda:{int(os.getenv('LOCAL_RANK', '0'))}"
-        agent_cfg.device = f"cuda:{int(os.getenv('LOCAL_RANK', '0'))}"
-
+        local_rank = int(os.getenv("LOCAL_RANK", "0"))
+        num_visible = torch.cuda.device_count()
+        world_size = int(os.getenv("WORLD_SIZE", "1"))
+        env_cfg.sim.device = f"cuda:{local_rank}" if num_visible >= world_size else "cuda:0"
+        agent_cfg.device = env_cfg.sim.device
+        if "cuda" in env_cfg.sim.device:
+            torch.cuda.set_device(env_cfg.sim.device)
         # set seed to have diversity in different threads
-        seed = agent_cfg.seed + int(os.getenv("LOCAL_RANK", "0"))
+        seed = agent_cfg.seed + local_rank
         env_cfg.seed = seed
         agent_cfg.seed = seed
         world_rank = int(os.getenv("RANK", "0"))
-        world_size = int(os.getenv("WORLD_SIZE", 1))
 
     # specify directory for logging experiments
     log_root_path = os.path.join("logs", "rsl_rl", agent_cfg.experiment_name)